Now server was able to receive msgs from client.
Added terminate for gen_server, The server was able to take the response from all the clients, it could also broadcast them to all the clients correspondingly. But the broadcast was not able to be printed on client. May be the client was able to get the msgs from broadcast through a spawned process bu was unable to print at the client side, since it was being blocked by input prompt from client at the same time.