Added slave process which should started in separate shell
added slave process which listens for broadcasts from client.Right now the slave process is having a fixed name slave which creates problem if we wantmultiple chat clients on the same server as this slave is registered globally(need to think of a solution for this). Initially this global slave process will be available to client only if the slave node is pinged(using net_kernel:connect_node(name@host)).