Little refactoring in broadcast
The server broadcast is transmitted without name, this name is dynamically added at the client to its own name which is incorrect. Now sending the broadcast with broadcaster name. On the client side name is not added any more. It just prints what comes in.
Although the messages are coming in binary and also added \n at the end.